+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -22,6 +22,7 @@
 * Azure Site Recovery support to select disk type at enabling protection.
 * Azure Site Recovery bug fix for recovery plan action edit.
 * Azure Backup SQL Restore support to accept filestream DBs.
+* Azure Backup updated SDK version.
 
 ## Version 2.0.1